requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
appdirs==1.4.4
asgiref==3.5.2
astroid==2.4.2
autopep8==1.5.4
black==20.8b1
boto3==1.14.31
botocore==1.17.31
cachelib==0.1
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cs50==5.0.4
distlib==0.3.1
dj-database-url==2.1.0
Django==3.2.1
django-admin-interface==0.14.0
django-admin-thumbnails==0.2.8
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-on-heroku==1.1.2
django-rest-passwordreset==1.2.1
django-storages==1.9.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.5.0
dnspython==2.2.1
docutils==0.15.2
drf-yasg==1.17.1
facebook-sdk==3.1.0
Faker==4.1.0
filelock==3.9.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-PyMongo==2.3.0
Flask-Session==0.3.2
future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==6.0.0
inflection==0.4.0
install==1.3.3
isort==5.6.4
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
lazy-object-proxy==1.4.3
macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.0
Pillow==8.0.1
platformdirs==2.5.2
protobuf==3.13.0
psycopg2-binary==2.9.3
psycopg2==2.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pymongo==4.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-twitter==3.5
pytz==2020.1
regex==2020.10.28
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.3.3
six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
SQLAlchemy==1.3.17
sqlparse==0.3.1
sweetify==2.3.1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
validate-email==1.3
virtualenv==20.2.1
Werkzeug==0.15.4
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.11.0
